Move the AST Builder to the correct project/namespace.
