+++ src/Core/Gibberish/Gibberish.csproj
@@ -64,6 +64,7 @@
     <Compile Include="AST\Block.cs" />
     <Compile Include="AST\Namespace.cs" />
     <Compile Include="AST\SpecificationSection.cs" />
+    <Compile Include="AST\_1_Bare\BasicAst.cs" />
     <Compile Include="Execution\City.cs" />
     <Compile Include="Compiler.cs" />
     <Compile Include="Execution\District.cs" />
